pyproject.toml
[tool.poetry]
name = "takehome"
version = "0.1.0"
description = ""
authors = ["Adam <adam@getsimple.rsvp>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.9.2"
dspy-ai = "^2.5.29"
together = "^1.3.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
